pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "product-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "fuzzywuzzy>=0.18.0",
    "lxml>=6.0.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
